edit: remove source file to avoid internal linting errors
